+++ src/main/java/com/sns/whisper/domain/user/application/LoginService.java
@@ -1,0 +1,7 @@
+package com.sns.whisper.domain.user.application;
+
+public interface LoginService {
+
+    void login(String userId, String password);
+
+}
