+++ src/test/java/com/sns/whisper/unit/user/application/UserServiceTest.java
@@ -4,13 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.sns.whisper.common.factory.UserFactory;
-import com.sns.whisper.common.mockapi.MockUserSessionManager;
 import com.sns.whisper.domain.user.application.UserService;
 import com.sns.whisper.domain.user.application.dto.request.UserSignUpServiceRequest;
 import com.sns.whisper.domain.user.application.dto.response.UserResponse;
@@ -21,7 +18,6 @@ import com.sns.whisper.domain.user.domain.profile.UserStatus;
 import com.sns.whisper.domain.user.domain.respository.ProfileStorage;
 import com.sns.whisper.domain.user.domain.respository.UserRepository;
 import com.sns.whisper.exception.user.DuplicatedUserIdException;
-import com.sns.whisper.exception.user.LoginFailException;
 import com.sns.whisper.exception.user.NotValidEmailFormatException;
 import com.sns.whisper.global.common.PasswordEncryptor;
 import java.time.LocalDate;
@@ -31,7 +27,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 
@@ -46,9 +41,6 @@ public class UserServiceTest {
 
     @Mock
     private UserRepository userRepository;
-
-    @Spy
-    private MockUserSessionManager sessionManager;
 
 
     @DisplayName("회원가입 요청을 받으면, 승인 대기 상태의 회원을 생성한다.")
@@ -107,71 +99,6 @@ public class UserServiceTest {
                 .isInstanceOf(DuplicatedUserIdException.class)
                 .hasFieldOrPropertyWithValue("httpStatus", HttpStatus.BAD_REQUEST)
                 .hasMessage("중복된 아이디입니다.");
-    }
-
-    @Test
-    @DisplayName("가입회원의 유효한 아이디와 비밀번호로 로그인할 수 있다.")
-    void login_ValidUser_Success() throws Exception {
-        //given
-        String userId = "userId123";
-        String password = "password1234";
-
-        User savedUser = UserFactory.createBasicUser(userId, password);
-        when(userRepository.findUserByUserId(userId)).thenReturn(Optional.of(savedUser));
-
-        //when
-        userService.login(userId, password);
-
-        //then
-        assertThat(sessionManager.extractUser("ACCESS_USER")).isEqualTo(userId);
-        assertThat(PasswordEncryptor.isMatch(password, savedUser.getPassword())).isTrue();
-        verify(userRepository).findUserByUserId(userId);
-    }
-
-    @Test
-    @DisplayName("가입하지 않은 회원은 로그인을 할 수 없다.")
-    void login_NotValidUserId_ExceptionThrown() throws Exception {
-        //given
-        String userId = "userId123";
-        String password = "password1234";
-
-        //when
-        assertThatCode(() -> userService.login(userId, password))
-                .isInstanceOf(
-                        LoginFailException.class)
-                .hasFieldOrPropertyWithValue(
-                        "httpStatus",
-                        HttpStatus.BAD_REQUEST)
-                .hasMessage("아이디와 비밀번호를 확인해주세요.");
-
-        verify(userRepository, times(1)).findUserByUserId(any(String.class));
-        verify(sessionManager, never()).saveUser(any(String.class));
-    }
-
-    @Test
-    @DisplayName("유효하지 않은 비밀번호를 입력할 경우 로그인을 할 수 없다.")
-    void login_NotValidPassword_ExceptionThrown() throws Exception {
-        //given
-        String userId = "userId123";
-        String password = "password1234";
-
-        String savedPassword = "testPassword1234";
-
-        User savedUser = UserFactory.createBasicUser(userId, savedPassword);
-
-        when(userRepository.findUserByUserId(userId)).thenReturn(Optional.of(savedUser));
-
-        //when
-        assertThatCode(() -> userService.login(userId, password))
-                .isInstanceOf(
-                        LoginFailException.class)
-                .hasFieldOrPropertyWithValue(
-                        "httpStatus",
-                        HttpStatus.BAD_REQUEST)
-                .hasMessage("아이디와 비밀번호를 확인해주세요.");
-
-        verify(userRepository, times(1)).findUserByUserId(any(String.class));
-        verify(sessionManager, never()).saveUser(any(String.class));
     }
 
     private UserSignUpServiceRequest createSignUpRequest(String email) {
